Group listening into sessions; drop redundant verb headers on single-verb feeds
On the dedicated Listening and Posting pages the per-item gerund header
("LISTENING TO", "POSTING") is redundant — the page already names the
verb. Add a `showVerb` prop to FeedItem so these pages render without the
verb column. Reply context isn't lost: PostCard's own inline reply badge
takes over when the verb column is hidden, and self-reply thread
continuations stay quiet.

Collapse the Listening feed into "listening sessions" the same way the
home feed already does — consecutive plays within a 2h gap merge into one
compact, expandable row instead of one row per track. The batching logic
(collapseListens) is extracted from Home into src/lib/listenSessions.js
so both pages share it.
